requirements.txt
Django
djangorestframework
django-cors-headers
torch
transformers
Pillow
pytesseract
layoutparser
psutil
pdfplumber